Move models to the root of the source directory
