+++ src/Models/TableRow.h
@@ -4,7 +4,7 @@
 #include <map>
 #include <string>
 
-#include "../Enums/DataTypes.h"
+#include "../Common/Enums/DataTypes.h"
 
 namespace FQL
 {
